+++ src/avt/Filters/avtSamplePointExtractor.C
@@ -1162,7 +1162,7 @@ avtSamplePointExtractor::ExtractHex(vtkHexahedron *hex, vtkDataSet *ds,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1180,7 +1180,7 @@ avtSamplePointExtractor::ExtractHex(vtkHexahedron *hex, vtkDataSet *ds,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
@@ -1252,7 +1252,7 @@ avtSamplePointExtractor::ExtractHex20(vtkQuadraticHexahedron *hex20,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1270,7 +1270,7 @@ avtSamplePointExtractor::ExtractHex20(vtkQuadraticHexahedron *hex20,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
@@ -1355,7 +1355,7 @@ avtSamplePointExtractor::ExtractWedge(vtkWedge *wedge, vtkDataSet *ds,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1373,7 +1373,7 @@ avtSamplePointExtractor::ExtractWedge(vtkWedge *wedge, vtkDataSet *ds,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
@@ -1458,7 +1458,7 @@ avtSamplePointExtractor::ExtractTet(vtkTetra *tet, vtkDataSet *ds,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1476,7 +1476,7 @@ avtSamplePointExtractor::ExtractTet(vtkTetra *tet, vtkDataSet *ds,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
@@ -1561,7 +1561,7 @@ avtSamplePointExtractor::ExtractPyramid(vtkPyramid *pyr, vtkDataSet *ds,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1579,7 +1579,7 @@ avtSamplePointExtractor::ExtractPyramid(vtkPyramid *pyr, vtkDataSet *ds,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
@@ -1661,7 +1661,7 @@ avtSamplePointExtractor::ExtractVoxel(vtkVoxel *voxel, vtkDataSet *ds,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1684,7 +1684,7 @@ avtSamplePointExtractor::ExtractVoxel(vtkVoxel *voxel, vtkDataSet *ds,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
@@ -1760,7 +1760,7 @@ avtSamplePointExtractor::ExtractTriangle(vtkTriangle *tri, vtkDataSet *ds,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1779,7 +1779,7 @@ avtSamplePointExtractor::ExtractTriangle(vtkTriangle *tri, vtkDataSet *ds,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
@@ -1867,7 +1867,7 @@ avtSamplePointExtractor::ExtractQuad(vtkQuad *quad, vtkDataSet *ds,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1885,7 +1885,7 @@ avtSamplePointExtractor::ExtractQuad(vtkQuad *quad, vtkDataSet *ds,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
@@ -1976,7 +1976,7 @@ avtSamplePointExtractor::ExtractPixel(vtkPixel *pixel, vtkDataSet *ds,
     //
     // Retrieve all of the zonal variables.
     //
-    int ncd = li.cellArrays.size();
+    int ncd = (int)li.cellArrays.size();
     for (v = 0 ; v < ncd ; v++)
     {
         if (li.cellDataIndex[v] < 0)
@@ -1994,7 +1994,7 @@ avtSamplePointExtractor::ExtractPixel(vtkPixel *pixel, vtkDataSet *ds,
     //
     // Retrieve all of the nodal variables.
     //
-    int npd = li.pointArrays.size();
+    int npd = (int)li.pointArrays.size();
     for (v = 0 ; v < npd ; v++)
     {
         if (li.pointDataIndex[v] < 0)
